+++ SAN/Brocade/FOS_PortErrShowInfos.ps1
@@ -25,6 +25,7 @@ function FOS_PortErrShowInfos {
         [string]$TD_Device_DeviceName,
         [string]$TD_Device_DeviceIP,
         [string]$TD_Device_PW,
+        [string]$TD_Device_SSHKeyPath,
         [Parameter(ValueFromPipeline)]
         [ValidateSet("yes","no")]
         [string]$TD_Export = "yes",
@@ -42,7 +43,7 @@ function FOS_PortErrShowInfos {
 
         if($TD_Device_ConnectionTyp -eq "ssh"){
             Write-Debug -Message "ssh |$(Get-Date)"
-            $FOS_MainInformation = ssh -i $($TD_tb_pathtokey.Text) $TD_Device_UserName@$TD_Device_DeviceIP "porterrshow"
+            $FOS_MainInformation = ssh -i $($TD_Device_SSHKeyPath) $TD_Device_UserName@$TD_Device_DeviceIP "porterrshow"
         }else {
             Write-Debug -Message "plink |$(Get-Date)"
             $FOS_MainInformation = plink $TD_Device_UserName@$TD_Device_DeviceIP -pw $TD_Device_PW -batch "porterrshow"