+++ TOOLFunc/SST_DeviceConnecCheck.ps1
@@ -7,6 +7,7 @@ function SST_DeviceConnecCheck {
         $TD_Selected_DeviceIPAddr,
         $TD_Selected_DeviceUserName,
         $TD_Selected_DevicePassword,
+        $TD_Selected_DeviceSSHFile,
         $TD_Selected_SVCorVF
     )
     
@@ -18,6 +19,7 @@ function SST_DeviceConnecCheck {
                 $TD_Selected_DeviceIPAddr
                 $TD_Selected_DeviceUserName
                 $TD_Selected_DevicePassword
+                $TD_Selected_DeviceSSHFile
                 $TD_UserInputCred = $TD_Selected_SVCorVF
              }
             "no" { 
@@ -25,6 +27,7 @@ function SST_DeviceConnecCheck {
                 $TD_Selected_DeviceIPAddr = $TD_TB_DeviceIPAddr.Text
                 $TD_Selected_DeviceUserName = $TD_TB_DeviceUserName.Text
                 $TD_Selected_DevicePassword = $TD_TB_DevicePassword.Password
+                $TD_Selected_DeviceSSHFile = $TD_TB_PathtoSSHKeyNotVisibil.Text="$($TD_ImportaddsshkeyObj.FileName)"
                 $TD_Selected_DeviceType = $TD_CB_DeviceType.Text
                 if($TD_CB_SVCorVF.IsChecked -and ($TD_Selected_DeviceType -eq "Storage")){$TD_UserInputCred = "SVC"};
                 if($TD_CB_SVCorVF.IsChecked -and ($TD_Selected_DeviceType -eq "SAN")){$TD_UserInputCred = "VF"};
@@ -39,7 +42,7 @@ function SST_DeviceConnecCheck {
 
         switch ($TD_Selected_DeviceType) {
             "Storage" { 
-                $TD_BasicDeviceInfos = IBM_BaseStorageInfos -TD_Device_ConnectionTyp $TD_Selected_DeviceConnectionType -TD_Device_DeviceIP $TD_Selected_DeviceIPAddr -TD_Device_UserName $TD_Selected_DeviceUserName -TD_Device_PW $([Net.NetworkCredential]::new('', $TD_Selected_DevicePassword).Password) -TD_Storage $TD_UserInputCred
+                $TD_BasicDeviceInfos = IBM_BaseStorageInfos -TD_Device_ConnectionTyp $TD_Selected_DeviceConnectionType -TD_Device_DeviceIP $TD_Selected_DeviceIPAddr -TD_Device_UserName $TD_Selected_DeviceUserName -TD_Device_PW $([Net.NetworkCredential]::new('', $TD_Selected_DevicePassword).Password) -TD_Device_SSHKeyPath $TD_Selected_DeviceSSHFile -TD_Storage $TD_UserInputCred
                 <# not the best check but try-catch do not work, i have to check why #>
                 if($TD_BasicDeviceInfos.count -gt 0){
                     $TD_BInfo = "" | Select-Object DeviceName,ProductDes,Prod_MTM,Code_Level
@@ -81,7 +84,7 @@ function SST_DeviceConnecCheck {
                 }
             }
             "SAN" { 
-                $TD_BasicDeviceInfos = FOS_BasicSwitchInfos -TD_Device_ConnectionTyp $TD_Selected_DeviceConnectionType -TD_Device_DeviceIP $TD_Selected_DeviceIPAddr -TD_Device_UserName $TD_Selected_DeviceUserName -TD_Device_PW $([Net.NetworkCredential]::new('', $TD_Selected_DevicePassword).Password)
+                $TD_BasicDeviceInfos = FOS_BasicSwitchInfos -TD_Device_ConnectionTyp $TD_Selected_DeviceConnectionType -TD_Device_DeviceIP $TD_Selected_DeviceIPAddr -TD_Device_UserName $TD_Selected_DeviceUserName -TD_Device_PW $([Net.NetworkCredential]::new('', $TD_Selected_DevicePassword).Password) -TD_Device_SSHKeyPath $TD_Selected_DeviceSSHFile 
                 if($TD_BasicDeviceInfos.count -gt 0){
                     $TD_BInfo = "" | Select-Object DeviceName,ProductDes,Prod_MTM,Code_Level
                     $TD_BInfo.DeviceName = $TD_BasicDeviceInfos.'Swicht Name'